refactor!: evict AppKit/AX and process plumbing from ProxySession
ProxySession imported AppKit (NSWorkspace Xcode discovery) and
ApplicationServices (the AX readiness probe) against the documented
module ownership. Now:

- ProcessRunner/OrderedPipeReader (generic process/pipe IO) move to
  ProxyCore, as do the UpstreamReadinessGate closure container and the
  XcodeTargetDiscovering protocol with its target type.
- XcodeReadinessProbe, the xcodeMCPBridge gate factory (plus a new
  liveDefault(config:clock:) policy wrapper), and
  LiveXcodeTargetDiscovery move to ProxyXcodeSupport, which already owns
  Xcode window inspection.
- The isDefaultXcrunMCPBridgeInvocation policy joins XcrunArguments in
  ProxyCore.
- RuntimeCoordinator's composition-root convenience init now takes the
  readiness gate and target discovery as injections; ProxyServer wires
  the live Xcode implementations. Tests construct stubs explicitly.
- ArchitectureTests now forbids AppKit/ApplicationServices imports in
  ProxySession so the boundary cannot silently regress.
