+++ Sources/ProxySession/DocumentationProvider.swift
@@ -1,5 +1,3 @@
-import AppKit
-import Darwin
 import Foundation
 import Logging
 import NIO
@@ -171,93 +169,6 @@ package enum DocumentationToolCatalog {
             return object["result"] as? [String: Any]
         }
     }
-}
-
-package struct DocumentationProviderTarget: Sendable, Equatable {
-    package let processID: pid_t
-    package let appPath: String
-    package let developerDir: String
-    package let mcpbridgePath: String
-
-    package init(
-        processID: pid_t,
-        appPath: String,
-        developerDir: String,
-        mcpbridgePath: String
-    ) {
-        self.processID = processID
-        self.appPath = appPath
-        self.developerDir = developerDir
-        self.mcpbridgePath = mcpbridgePath
-    }
-}
-
-package protocol XcodeTargetDiscovering: Sendable {
-    func runningXcodeTargets() -> [DocumentationProviderTarget]
-}
-
-package struct LiveXcodeTargetDiscovery: XcodeTargetDiscovering {
-    package init() {}
-
-    package func runningXcodeTargets() -> [DocumentationProviderTarget] {
-        var targetsByPID: [pid_t: DocumentationProviderTarget] = [:]
-
-        for application in NSWorkspace.shared.runningApplications {
-            guard application.bundleIdentifier == "com.apple.dt.Xcode",
-                  application.isTerminated == false,
-                  let bundlePath = application.bundleURL?.path else {
-                continue
-            }
-            if let target = Self.target(processID: application.processIdentifier, appPath: bundlePath) {
-                targetsByPID[target.processID] = target
-            }
-        }
-
-        for processID in ProcessEnumeration.processIDs(named: "Xcode") {
-            if targetsByPID[processID] != nil {
-                continue
-            }
-            guard let processPath = ProcessEnumeration.executablePath(of: processID),
-                  let appPath = Self.appPath(fromExecutablePath: processPath),
-                  let target = Self.target(processID: processID, appPath: appPath) else {
-                continue
-            }
-            targetsByPID[target.processID] = target
-        }
-
-        return targetsByPID.values.sorted { lhs, rhs in
-            if lhs.appPath == rhs.appPath {
-                return lhs.processID < rhs.processID
-            }
-            return lhs.appPath < rhs.appPath
-        }
-    }
-
-    private static func target(processID: pid_t, appPath: String) -> DocumentationProviderTarget? {
-        let developerDir = URL(fileURLWithPath: appPath)
-            .appendingPathComponent("Contents/Developer")
-            .path
-        let mcpbridgePath = URL(fileURLWithPath: developerDir)
-            .appendingPathComponent("usr/bin/mcpbridge")
-            .path
-        guard FileManager.default.isExecutableFile(atPath: mcpbridgePath) else {
-            return nil
-        }
-        return DocumentationProviderTarget(
-            processID: processID,
-            appPath: appPath,
-            developerDir: developerDir,
-            mcpbridgePath: mcpbridgePath
-        )
-    }
-
-    private static func appPath(fromExecutablePath path: String) -> String? {
-        guard let range = path.range(of: "/Contents/MacOS/Xcode") else {
-            return nil
-        }
-        return String(path[..<range.lowerBound])
-    }
-
 }
 
 package protocol DocumentationProviderSessionMaking: Sendable {
@@ -505,7 +416,7 @@ package actor DocumentationProviderManager: DocumentationProviderManaging {
     private var isShutdown = false
 
     package init(
-        discovery: any XcodeTargetDiscovering = LiveXcodeTargetDiscovery(),
+        discovery: any XcodeTargetDiscovering,
         sessionFactory: any DocumentationProviderSessionMaking = LiveDocumentationProviderSessionFactory(),
         providerSelectionTimeout: TimeAmount? = .seconds(30),
         pinnedProcessID: pid_t? = nil,
