+++ Sources/ProxySession/Upstream/UpstreamReadinessGate.swift
@@ -1,244 +1,7 @@
 import Foundation
-import ApplicationServices
 import Logging
 import NIOConcurrencyHelpers
-
-package struct UpstreamReadinessGate: Sendable {
-    package let isEnabled: Bool
-    package let targetName: String
-    package let pollIntervalNanoseconds: UInt64
-    package let progressLogIntervalNanoseconds: UInt64
-    package let launchRetryIntervalNanoseconds: UInt64
-    package let initialRetryBackoffNanoseconds: UInt64
-    package let maxRetryBackoffNanoseconds: UInt64
-    package let uptimeNanoseconds: @Sendable () -> UInt64
-    package let sleepNanoseconds: @Sendable (UInt64) async -> Void
-    package let isAvailable: (@Sendable () async -> Bool)?
-    package let launchIfUnavailable: (@Sendable () async -> Bool)?
-    package let isReady: @Sendable () async -> Bool
-
-    package init(
-        isEnabled: Bool,
-        targetName: String,
-        pollIntervalNanoseconds: UInt64,
-        progressLogIntervalNanoseconds: UInt64,
-        launchRetryIntervalNanoseconds: UInt64,
-        initialRetryBackoffNanoseconds: UInt64,
-        maxRetryBackoffNanoseconds: UInt64,
-        uptimeNanoseconds: @escaping @Sendable () -> UInt64,
-        sleepNanoseconds: @escaping @Sendable (UInt64) async -> Void,
-        isAvailable: (@Sendable () async -> Bool)? = nil,
-        launchIfUnavailable: (@Sendable () async -> Bool)? = nil,
-        isReady: @escaping @Sendable () async -> Bool
-    ) {
-        self.isEnabled = isEnabled
-        self.targetName = targetName
-        self.pollIntervalNanoseconds = pollIntervalNanoseconds
-        self.progressLogIntervalNanoseconds = progressLogIntervalNanoseconds
-        self.launchRetryIntervalNanoseconds = launchRetryIntervalNanoseconds
-        self.initialRetryBackoffNanoseconds = initialRetryBackoffNanoseconds
-        self.maxRetryBackoffNanoseconds = maxRetryBackoffNanoseconds
-        self.uptimeNanoseconds = uptimeNanoseconds
-        self.sleepNanoseconds = sleepNanoseconds
-        self.isAvailable = isAvailable
-        self.launchIfUnavailable = launchIfUnavailable
-        self.isReady = isReady
-    }
-
-    package static func alwaysReady(
-        uptimeNanoseconds: @escaping @Sendable () -> UInt64 = {
-            DispatchTime.now().uptimeNanoseconds
-        }
-    ) -> Self {
-        Self(
-            isEnabled: false,
-            targetName: "upstream",
-            pollIntervalNanoseconds: 0,
-            progressLogIntervalNanoseconds: 0,
-            launchRetryIntervalNanoseconds: 0,
-            initialRetryBackoffNanoseconds: 0,
-            maxRetryBackoffNanoseconds: 0,
-            uptimeNanoseconds: uptimeNanoseconds,
-            sleepNanoseconds: { _ in },
-            isAvailable: { true },
-            launchIfUnavailable: nil,
-            isReady: { true }
-        )
-    }
-
-    package static func xcodeMCPBridge(
-        uptimeNanoseconds: @escaping @Sendable () -> UInt64,
-        sleepNanoseconds: @escaping @Sendable (UInt64) async -> Void,
-        runProcess: @escaping @Sendable (ProcessRequest) async throws -> ProcessOutput
-    ) -> Self {
-        Self(
-            isEnabled: true,
-            targetName: "mcpbridge",
-            pollIntervalNanoseconds: 1_000_000_000,
-            progressLogIntervalNanoseconds: 5_000_000_000,
-            launchRetryIntervalNanoseconds: 5_000_000_000,
-            initialRetryBackoffNanoseconds: 1_000_000_000,
-            maxRetryBackoffNanoseconds: 8_000_000_000,
-            uptimeNanoseconds: uptimeNanoseconds,
-            sleepNanoseconds: sleepNanoseconds,
-            isAvailable: {
-                let output = try? await runProcess(
-                    ProcessRequest(
-                        label: "detect-xcode-process",
-                        executablePath: "/usr/bin/pgrep",
-                        arguments: ["-x", "Xcode"],
-                        input: nil
-                    )
-                )
-                guard let output else { return false }
-                return XcodeReadinessProbe.processIDs(fromPGrepOutput: output).isEmpty == false
-            },
-            launchIfUnavailable: {
-                let output = try? await runProcess(
-                    ProcessRequest(
-                        label: "launch-xcode",
-                        executablePath: "/usr/bin/open",
-                        arguments: ["-a", "Xcode"],
-                        input: nil
-                    )
-                )
-                return output?.terminationStatus == 0
-            },
-            isReady: {
-                let output = try? await runProcess(
-                    ProcessRequest(
-                        label: "detect-xcode-process",
-                        executablePath: "/usr/bin/pgrep",
-                        arguments: ["-x", "Xcode"],
-                        input: nil
-                    )
-                )
-                guard let output else { return false }
-                let processIDs = XcodeReadinessProbe.processIDs(fromPGrepOutput: output)
-                return XcodeReadinessProbe.isReady(
-                    xcodeProcessIDs: processIDs,
-                    windows: XcodeReadinessProbe.visibleWindowSnapshots(
-                        xcodeProcessIDs: processIDs
-                    )
-                )
-            }
-        )
-    }
-}
-
-package enum XcodeReadinessProbe {
-    package struct WindowSnapshot: Equatable, Sendable {
-        package let ownerPID: pid_t
-        package let title: String
-        package let layer: Int
-        package let alpha: Double
-
-        package init(ownerPID: pid_t, title: String, layer: Int, alpha: Double) {
-            self.ownerPID = ownerPID
-            self.title = title
-            self.layer = layer
-            self.alpha = alpha
-        }
-    }
-
-    package static func processIDs(fromPGrepOutput output: ProcessOutput) -> Set<pid_t> {
-        guard output.terminationStatus == 0 else { return [] }
-        return Set(
-            output.stdout
-                .split(whereSeparator: \.isNewline)
-                .compactMap { pid_t($0.trimmingCharacters(in: .whitespacesAndNewlines)) }
-        )
-    }
-
-    package static func hasReadyWorkspaceWindow(
-        xcodeProcessIDs: Set<pid_t>,
-        windows: [WindowSnapshot]
-    ) -> Bool {
-        guard xcodeProcessIDs.isEmpty == false else { return false }
-        return windows.contains { window in
-            xcodeProcessIDs.contains(window.ownerPID)
-                && window.layer == 0
-                && window.alpha > 0
-                && isReadyWorkspaceTitle(window.title)
-        }
-    }
-
-    package static func isReady(
-        xcodeProcessIDs: Set<pid_t>,
-        windows: [WindowSnapshot]?
-    ) -> Bool {
-        guard xcodeProcessIDs.isEmpty == false else { return false }
-        guard let windows else {
-            return true
-        }
-        return hasReadyWorkspaceWindow(xcodeProcessIDs: xcodeProcessIDs, windows: windows)
-    }
-
-    package static func visibleWindowSnapshots(xcodeProcessIDs: Set<pid_t>) -> [WindowSnapshot]? {
-        guard AXIsProcessTrusted() else {
-            return nil
-        }
-        return accessibilityWindowSnapshots(xcodeProcessIDs: xcodeProcessIDs)
-    }
-
-    private static func accessibilityWindowSnapshots(
-        xcodeProcessIDs: Set<pid_t>
-    ) -> [WindowSnapshot] {
-        xcodeProcessIDs.flatMap { processID -> [WindowSnapshot] in
-            let app = AXUIElementCreateApplication(processID)
-            var rawWindows: CFTypeRef?
-            let error = unsafe AXUIElementCopyAttributeValue(
-                app,
-                kAXWindowsAttribute as CFString,
-                &rawWindows
-            )
-            guard error == .success, let windows = rawWindows as? [AXUIElement] else {
-                return []
-            }
-            return windows.map { window in
-                let minimized = accessibilityBoolAttribute(
-                    window,
-                    attribute: kAXMinimizedAttribute
-                ) ?? false
-                return WindowSnapshot(
-                    ownerPID: processID,
-                    title: accessibilityStringAttribute(window, attribute: kAXTitleAttribute) ?? "",
-                    layer: 0,
-                    alpha: minimized ? 0 : 1
-                )
-            }
-        }
-    }
-
-    private static func isReadyWorkspaceTitle(_ title: String) -> Bool {
-        let trimmed = title.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard trimmed.isEmpty == false else { return false }
-        guard trimmed.localizedCaseInsensitiveCompare("Welcome to Xcode") != .orderedSame else {
-            return false
-        }
-        return true
-    }
-
-    private static func accessibilityStringAttribute(
-        _ element: AXUIElement,
-        attribute: String
-    ) -> String? {
-        var value: CFTypeRef?
-        let error = unsafe AXUIElementCopyAttributeValue(element, attribute as CFString, &value)
-        guard error == .success else { return nil }
-        return value as? String
-    }
-
-    private static func accessibilityBoolAttribute(
-        _ element: AXUIElement,
-        attribute: String
-    ) -> Bool? {
-        var value: CFTypeRef?
-        let error = unsafe AXUIElementCopyAttributeValue(element, attribute as CFString, &value)
-        guard error == .success else { return nil }
-        return value as? Bool
-    }
-}
+import ProxyCore
 
 package final class UpstreamReadinessWaiterToken: Sendable {
     private let isCancelledBox = NIOLockedValueBox(false)
